(PDB-3702) Update behavior in acceptance tests
Before this commit tests would always default to using the latest release
of pdb regardless of the branch being tested against. This could cause
situations where we would install a later version of pdb and then try to
downgrade to an older version in the upgrade step. This adds branch specific
behavior and accounts for newly supported platforms not having 4.2.3.8 builds.