(DO NOT MERGE) Demonstrate queries using 2 hikari connections per query
This commit should not be merged. It uses the PDBReadPool metrics to
show that PDB queries are grabbing two connections per query. I believe
this is due to the double wrapping of with-transacted-connection which
eventually calls with-db-connection. It appears that nesting these
forms causes the query to use 2 connections from the hikari pool per
query.